brain/requirements.txt
awsiotsdk==1.7.1
pyserial==3.5
redis==4.1